requirements.txt
boto3==1.36.23
pyjwt==2.3.0
.
